+++ lib/lib/state/time.dart
@@ -4,12 +4,17 @@ import 'package:flutter/material.dart';
 
 /// Model wrapper class [Time] for [TimeOfDay]
 class Time extends TimeOfDay {
+  int second = 0;
+
   /// Constructor for the class
-  const Time(int hour, int minute) : super(hour: hour, minute: minute);
+  Time({required int hour, required int minute, int? second})
+      : super(hour: hour, minute: minute) {
+    this.second = second ?? 0;
+  }
 
   /// Get [Time] instance from [TimeOfDay]
-  factory Time.fromTimeOfDay(TimeOfDay time) {
-    return Time(time.hour, time.minute);
+  factory Time.fromTimeOfDay(TimeOfDay time, int? secondVal) {
+    return Time(hour: time.hour, minute: time.minute, second: secondVal);
   }
 
   /// Get [TimeOfDay] instance from [Time]
@@ -24,8 +29,11 @@ class Time extends TimeOfDay {
 
   /// Overide [TimeOfDay.replacing]
   @override
-  Time replacing({int? hour, int? minute}) {
-    return Time.fromTimeOfDay(super.replacing(hour: hour, minute: minute));
+  Time replacing({int? hour, int? minute, int? second}) {
+    return Time.fromTimeOfDay(
+      super.replacing(hour: hour, minute: minute),
+      second ?? this.second,
+    );
   }
 
   /// Helper for toggling period